Изменил методы рисования дуг с целочисленных градусов на дробные радианы для удобства использования в стратегии

+++ clients/java/VisualClient.java
@@ -120,19 +120,19 @@ public class VisualClient {
     }
 
     /**
-     * draw a arc with center at (centerX, centerY), radius radius and angle arcAngle, started from startAngle with color color
+     * draw a arc with center at (centerX, centerY), radius radius and angle arcAngle, started from startAngle with color color, angles in radians
      */
-    public void arc(double centerX, double centerY, double radius, int startAngle, int arcAngle, Color color) {
+    public void arc(double centerX, double centerY, double radius, double startAngle, double arcAngle, Color color) {
         Formatter f = new Formatter();
-        sendCommand(f.format("arc %1.1f %1.1f %1.1f %d %d %1.1f %1.1f %1.1f", centerX, centerY, radius, startAngle, arcAngle, (float) color.getRed()/255, (float) color.getGreen()/255, (float) color.getBlue()/255).toString());
+        sendCommand(f.format("arc %1.1f %1.1f %1.1f %1.1f %1.1f %1.1f %1.1f %1.1f", centerX, centerY, radius, startAngle, arcAngle, (float) color.getRed()/255, (float) color.getGreen()/255, (float) color.getBlue()/255).toString());
     }
 
     /**
-     * draw a filled arc with center at (centerX, centerY), radius radius and angle arcAngle, started from startAngle with color color
+     * draw a filled arc with center at (centerX, centerY), radius radius and angle arcAngle, started from startAngle with color color, angles in radians
      */
-    public void fillArc(double centerX, double centerY, double radius, int startAngle, int arcAngle, Color color) {
+    public void fillArc(double centerX, double centerY, double radius, double startAngle, double arcAngle, Color color) {
         Formatter f = new Formatter();
-        sendCommand(f.format("fill_arc %1.1f %1.1f %1.1f %d %d %1.1f %1.1f %1.1f", centerX, centerY, radius, startAngle, arcAngle, (float) color.getRed()/255, (float) color.getGreen()/255, (float) color.getBlue()/255).toString());
+        sendCommand(f.format("fill_arc %1.1f %1.1f %1.1f %1.1f %1.1f %1.1f %1.1f %1.1f", centerX, centerY, radius, startAngle, arcAngle, (float) color.getRed()/255, (float) color.getGreen()/255, (float) color.getBlue()/255).toString());
     }
 
     /**
